Move worker related files into its own directory
