+++ arangod/Pregel/CMakeLists.txt
@@ -2,7 +2,6 @@ add_library(arango_pregel STATIC
   AggregatorHandler.cpp
   AlgoRegistry.cpp
   ExecutionNumber.cpp
-  GraphStore.cpp
   IncomingCache.cpp
   IndexHelpers.cpp
   OutgoingCache.cpp
@@ -10,9 +9,7 @@ add_library(arango_pregel STATIC
   PregelMetrics.cpp
   Reports.cpp
   Reports.h
-  Utils.cpp
-  Worker.cpp
-  WorkerConfig.cpp)
+  Utils.cpp)
 
 target_link_libraries(arango_pregel
   arango
@@ -27,3 +24,4 @@ target_include_directories(arango_pregel PRIVATE
 add_subdirectory(Algos)
 add_subdirectory(Conductor)
 add_subdirectory(REST)
+add_subdirectory(Worker)
