+++ arangod/Pregel/Worker/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_sources(arango_pregel PRIVATE
+  GraphStore.cpp
+  WorkerConfig.cpp
+  Worker.cpp)
