+++ arangod/V8Server/v8-pregel.cpp
@@ -32,7 +32,7 @@
 #include "Pregel/AggregatorHandler.h"
 #include "Pregel/Conductor/Conductor.h"
 #include "Pregel/PregelFeature.h"
-#include "Pregel/Worker.h"
+#include "Pregel/Worker/Worker.h"
 
 #include "V8/v8-conv.h"
 #include "V8/v8-globals.h"
